Failed attempt to switch pyvenv->virtualenvwrapper in eshell.
I think the swap from `pyvenv` to `virtualenvwrapper` is a good move that I
need to spend more time on.

My eshell changes on the other hand were no good. I broke the buffer nature of
eshell so couldn't move around the buffer with the cursor keys (although did
have the current virtualenv in the prompt). Also the eshell virtualenv still
doesn't behave in the same way as the stock terminal for some reason.