Tweaked shr luminance value to make some work emails more visible.
Get some work emails in mu4e that appear to be double nested in the font-faces,
which means that they come out as white on white, instead of white on black, or
black on white.

Tweaked value so that all emails are "okay-ish" instead of some "great" and
others white on white. Not a great solution, but good enough for now.