Disable `pipenv`. Starting to have some weird issues with it.
Not sure if it was me screwing up an old python hook to use `lsp-mode` instead
of `lsp` (which I've finally fixed), but this has been playing up recently.

Generally I'm now on the fence on using `pipenv` since it seems to be losing
support against `poetry`. Use it locally still, but moving back to using
`venv`/`pyvenv` more.